chore: Phuc - add location json file
