+++ src/Core/Files/Statics/css/swagger-ui.css
